+++ src/presenter/main-presenter.js
@@ -1,20 +1,22 @@
 import { render } from '../render';
-import EditFormView from '../view/edit-form-view';
+// import EditFormView from '../view/edit-form-view';
 import FiltersView from '../view/filters-view';
 import PointView from '../view/point-view';
 import PointsListView from '../view/points-list-view';
 import SortView from '../view/sort-view';
-import { BLANK_POINT } from '../const';
+// import { BLANK_POINT } from '../const';
 
 // $======================== MainPresenter ========================$ //
 
 export default class MainPresenter {
   pointsListElement = new PointsListView();
 
-  constructor({ pointsContainer, filtersContainer, pointsModel }) {
+  constructor({ pointsContainer, filtersContainer, pointsModel, destinationsModel, offersModel }) {
     this.pointsContainer = pointsContainer;
     this.filtersContainer = filtersContainer;
     this.pointsModel = pointsModel;
+    this.destinationsModel = destinationsModel;
+    this.offersModel = offersModel;
   }
 
   init() {
@@ -28,8 +30,8 @@ export default class MainPresenter {
     for (let i = 0; i < this.points.length; i++) {
       const point = new PointView({
         point: this.points[i],
-        offers: [...this.pointsModel.getOffersById(this.points[i].type, this.points[i].offers)],
-        destination: this.pointsModel.getDestinationById(this.points[i].destination).name
+        offers: [...this.offersModel.getOffersById(this.points[i].type, this.points[i].offers)],
+        destination: this.destinationsModel.getDestinationById(this.points[i].destination)
       });
       render(point, this.pointsListElement.getElement());
     }
