+++ src/view/point-view.js
@@ -31,7 +31,7 @@ const createPointTemplate = ({ point, offers, destination }) => {
           <img class="event__type-icon" width="42" height="42" src="img/icons/${type}.png" alt="Event type icon">
         </div>
 
-        <h3 class="event__title">${type} ${destination}</h3>
+        <h3 class="event__title">${type} ${destination.name}</h3>
 
         <div class="event__schedule">
           <p class="event__time">
